pyproject.toml
[project]
name = "runpod-flash"
version = "0.23.1"
description = "A Python library for distributed inference and serving of machine learning models"
authors = [
    { name = "Runpod", email = "engineer@runpod.io" },
]
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10,<3.15"

dependencies = [
    "cloudpickle>=3.1.1",
    "runpod",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "typer>=0.12.0",
    "questionary>=2.0.0",
    "pathspec>=0.11.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.11.9",
    "mcp>=1.0.0",
]
test = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.6.1",
    "twine>=6.1.0",
]

[project.scripts]
flash = "runpod_flash.cli.main:app"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
runpod_flash = [
    "cli/utils/skeleton_template/**/*",
    "cli/utils/skeleton_template/**/.*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--cov=runpod_flash",
    "--cov-report=term-missing",
    "--cov-fail-under=65"
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "serial: Tests that must run serially (not parallelized)"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::pytest.PytestDeprecationWarning",
    "ignore::pytest.PytestUnknownMarkWarning"
]

[tool.mypy]
# Basic configuration
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, can be stricter later
disallow_incomplete_defs = false
check_untyped_defs = true

# Import discovery
mypy_path = "."
namespace_packages = true

# Error output
show_error_codes = true
show_column_numbers = true
pretty = true

# Exclude directories
exclude = [
    "tests/",  # Start by excluding tests, can add later
]

# Per-module options
[[tool.mypy.overrides]]
module = [
    "runpod.*",
    "cloudpickle.*",
]
ignore_missing_imports = true

[tool.coverage.run]
parallel = true
branch = false
source = ["runpod_flash"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.paths]
source = [
    "src/runpod_flash",
    "*/site-packages/runpod_flash",
]

[tool.uv.sources]
runpod = { git = "https://github.com/runpod/runpod-python", rev = "main" }
